Complete user object and some DOM functionalities
A user object has been created and completed. It
contains all the projects and has a name (the
user name). This is useful to group all the
projects and todos.

Some progress has been made on the webpage as
well. By default the page loads all the tasks from
all the projects from the user. It also shows
the different projects in the sidebar.

A lot more is needed to complete the page. All the
siebar buttons need their own working
functionality. The styling of the page has not
been done yet either.