Complete 'Add Project' form and logic
The 'Add Project' button now creates a form to Add
a new project. It has an input text field for the
project name and two buttons: one to add and one
to cancel.

Also deleted functions' need to have the user
object passed as a parameter, it is now imported
in the modules where it's needed.

Added a 'Add new task' form on each project's page.
There is no logic attached to it yet (next commit).
It will enable the user to add a task to the
project in question.